+++ include/lexer/buffer.hpp
@@ -8,7 +8,7 @@ struct Buffer {
 	std::deque<char> pushback_buffer;
 
 	const char* source = NULL;
-	unsigned int line = 1, col = 1;
+	size_t line = 1, col = 1;
 
 	Buffer (const char* filename);
 	~Buffer ();
@@ -17,13 +17,13 @@ struct Buffer {
 	char next ();
 	void pushback (char c);
 	bool hasNext ();
-	char peek (unsigned int offset = 0);
+	char peek (size_t offset = 0);
 	bool is_next (char c);
 	bool is_next (const char* expected);
-	void skip (unsigned int count = 1);
+	void skip (size_t count = 1);
 	void skipAny (const char* chars);
 	void skipUntil (const char* stopper);
 	void printLocation ();
-	void fillPushbackBuffer (unsigned int limit);
+	void fillPushbackBuffer (size_t limit);
 	void handleLineCol (char character);
 };
